requirements.txt
supabase
httpx
python-dotenv
pytest
